+++ CHANGES.rst
@@ -7,6 +7,7 @@ Release 0.1.3(in development)
 
 Incompatible changes
 --------------------
+* Changed license from BSD to MIT.
 
 Features added
 --------------
@@ -16,6 +17,7 @@ Bugs fixed
 
 Documentation
 -------------
+* Added more badges.
 
 
 Release 0.1.2